\name{knots}
\alias{knots}
\alias{k3_1}
\alias{k3_1a}
\alias{k4_1}
\alias{k4_1a}
\alias{k5_1}
\alias{k5_2}
\alias{k6_1}
\alias{k6_2}
\alias{k6_3}
\alias{k7_1}
\alias{k7_2}
\alias{k7_3}
\alias{k7_4}
\alias{k7_5}
\alias{k7_6}
\alias{k7_7}
\alias{k7_7a}
\alias{k8_1}
\alias{k8_2}
\alias{k8_3}
\alias{k8_3_90deg_crossing}
\alias{k8_4}
\alias{k8_4a}
\alias{k8_5}
\alias{k8_6}
\alias{k8_7}
\alias{k8_8}
\alias{k8_9}
\alias{k8_10}
\alias{k8_11}
\alias{k8_12}
\alias{k8_13}
\alias{k8_14}
\alias{k8_15}
\alias{k8_16}
\alias{k8_17}
\alias{k8_18}
\alias{k8_19}
\alias{k8_19a}
\alias{k8_19b}
\alias{k8_20}
\alias{k8_21}
\alias{k9_1}
\alias{k9_2}
\alias{k9_3}
\alias{k9_4}
\alias{k9_5}
\alias{k9_6}
\alias{k9_7}
\alias{k9_8}
\alias{k9_9}
\alias{k9_10}
\alias{k9_11}
\alias{k9_12}
\alias{k9_13}
\alias{k9_14}
\alias{k9_15}
\alias{k9_16}
\alias{k9_17}
\alias{k9_18}
\alias{k9_19}
\alias{k9_20}
\alias{k9_21}
\alias{k9_22}
\alias{k9_23}
\alias{k9_23a}
\alias{k9_24}
\alias{k9_25}
\alias{k9_26}
\alias{k9_27}
\alias{k9_28}
\alias{k9_29}
\alias{k9_30}
\alias{k9_31}
\alias{k9_32}
\alias{k9_33}
\alias{k9_34}
\alias{k9_35}
\alias{k9_36}
\alias{k9_37}
\alias{k9_38}
\alias{k9_39}
\alias{k9_40}
\alias{k9_41}
\alias{k9_42}
\alias{k9_43}
\alias{k9_44}
\alias{k9_45}
\alias{k9_46}
\alias{k9_47}
\alias{k9_48}
\alias{k9_49}
\alias{k10_1}
\alias{k10_47}
\alias{k10_61}
\alias{k10_61a}
\alias{k10_123}
\alias{k11a179}
\alias{k11a361}
\alias{k11a1}
\alias{k11n157}
\alias{k11n157_morenodes}
\alias{k11n22}
\alias{k12n_0242}
\alias{k12n_0411}
\alias{k12a1202}
\alias{k12n838}
\alias{celtic3}
\alias{k_infinity}
\alias{fourloops}
\alias{fiveloops}
\alias{flower}
\alias{hexknot}
\alias{hexknot2}
\alias{hexknot3}
\alias{longthin}
\alias{ochiai}
\alias{ornamental20}
\alias{perko_A}
\alias{perko_B}
\alias{pretzel_2_3_7}
\alias{pretzel_7_3_7}
\alias{pretzel_p3_p5_p7_m3_m5}
\alias{product_knot}
\alias{reefknot}
\alias{amphichiral15}
\alias{satellite}
\alias{sum_31_41}
\alias{three_figure_eights}
\alias{trefoil_of_trefoils}
\alias{triloop}
\alias{unknot}
\alias{D16}
\alias{T20}
\alias{k8_11_90deg_crossing}
\alias{k8_5_90deg_crossing}
\alias{k8_6_90deg_crossing}
\alias{pretzel_2_3_7_90deg_crossing}
\alias{pretzel_7_3_7_90deg_crossing}
\docType{data}
\title{Optimized knots}
\description{A variety of knots with optimized forms}
\details{

  A selection of knots that have been optimized for visual appearance.
  The list makes no claims for completeness; the examples are intended
  to show the abilities of the package.

  Knots with names like \code{k7_3} use the naming scheme of Rolfsen.

  Knots with names like \code{k11n157} follow the nomenclature of the
  Hoste-Thistlethwaite table;  \sQuote{a}  means \sQuote{alternating}
  and \sQuote{n} means \sQuote{nonalternating}.
  
  Knot \code{k12a_614} is drawn from the \dQuote{Table of Knot
    Invariants} by Livingstone and Cha.
  
  Knot \code{amphichiral15} is the unique amphichiral knot with crossing
  number 15, due to Hoste, Thistlethwaite, and Weeks.

  Knots \code{k12n_0411} and \code{k11a203} show that partial symmetry
  may be enforced.

  Knot \code{k8_18} is an exceptional knot.

  Knot \code{pretzel_p3_p5_p7_m3_m5} is drawn from a knot appearing in
  Bryant 2016.  The notation specifies the sense (\sQuote{p} for plus
  and \sQuote{m} for minus) of the twists.

  Knot \code{T20} is a \dQuote{remarkable 20-crossing tangle}; see
  references

  Knots \code{k12a1202} and \code{k12n838} are named following Lamm.

  As of version 1.0-4, the complete list of knots is:

\code{k10_1}, \code{k10_123}, \code{k10_47}, \code{k10_61},
\code{k12a1202}, \code{k12n838}, \code{k3_1}, \code{k3_1a}, \code{k4_1},
\code{k4_1a}, \code{k5_1}, \code{k5_2}, \code{k6_1}, \code{k6_2},
\code{k6_3}, \code{k7_1}, \code{k7_2}, \code{k7_3}, \code{k7_4},
\code{k7_5}, \code{k7_6}, \code{k7_7}, \code{k7_7a}, \code{k8_1},
\code{k8_10}, \code{k8_11}, \code{k8_12}, \code{k8_13}, \code{k8_14},
\code{k8_15}, \code{k8_16}, \code{k8_17}, \code{k8_18}, \code{k8_19},
\code{k8_19a}, \code{k8_19b}, \code{k8_2}, \code{k8_20}, \code{k8_21},
\code{k8_3}, \code{k8_3_90deg_crossing},
\code{k8_4}, \code{k8_4a}, \code{k8_5}, \code{k8_6},
\code{k8_7}, \code{k8_8}, \code{k8_9}, \code{k9_1}, \code{k9_10},
\code{k9_11}, \code{k9_12}, \code{k9_13}, \code{k9_14}, \code{k9_15},
\code{k9_16}, \code{k9_17}, \code{k9_18}, \code{k9_19}, \code{k9_2},
\code{k9_20}, \code{k9_21}, \code{k9_22}, \code{k9_23}, \code{k9_23a},
\code{k9_24}, \code{k9_25}, \code{k9_26}, \code{k9_27}, \code{k9_28},
\code{k9_29}, \code{k9_3}, \code{k9_30}, \code{k9_31}, \code{k9_32},
\code{k9_33}, \code{k9_34}, \code{k9_35}, \code{k9_36}, \code{k9_37},
\code{k9_38}, \code{k9_39}, \code{k9_4}, \code{k9_40}, \code{k9_41},
\code{k9_42}, \code{k9_43}, \code{k9_44}, \code{k9_45}, \code{k9_46},
\code{k9_47}, \code{k9_48}, \code{k9_49}, \code{k9_5}, \code{k9_6},
\code{k9_7}, \code{k9_8}, \code{k9_9}, \code{D16}, \code{T20},
\code{amphichiral15}, \code{celtic3}, \code{fiveloops}, \code{flower},
\code{fourloops}, \code{hexknot}, \code{hexknot2}, \code{hexknot3},
\code{k_infinity}, \code{k11a1}, \code{k11a179}, \code{k11a361},
\code{k11n157}, \code{k11n157_morenodes}, \code{k11n22},
\code{k12n_0242}, \code{k12n_0411}, \code{longthin}, \code{ochiai},
\code{ornamental20}, \code{perko_A}, \code{perko_B},
\code{pretzel_2_3_7}, \code{pretzel_7_3_7},
\code{pretzel_p3_p5_p7_m3_m5}, \code{reefknot}, \code{satellite},
\code{sum_31_41}, \code{three_figure_eights},
\code{trefoil_of_trefoils}, \code{triloop}, \code{unknot}

}
\references{
  \itemize{
    \item K. A. Bryant, 2016. \emph{Slice implies mutant-ribbon for odd,
      5-stranded pretzel knots}, \code{arXiv:1511.07009v2}
    \item S. Eliahou and J. Fromentin 2017. \dQuote{A remarkable 20-crossing
      tangle}.  Arxiv, \url{https://arxiv.org/abs/1610.05560v2}
  }
}

\examples{
knotplot(k3_1)
## maybe str(k3_1) ; plot(k3_1) ...
}
\keyword{datasets}
